use long for ids

+++ domain/src/main/java/com/onlydust/marketplace/indexer/domain/model/clean/CodeReview.java
@@ -1,4 +1,4 @@
 package com.onlydust.marketplace.indexer.domain.model.clean;
 
-public record CodeReview(Integer id, User author) {
+public record CodeReview(Long id, User author) {
 }
