+++ domain/src/main/java/com/onlydust/marketplace/indexer/domain/model/clean/PullRequest.java
@@ -2,6 +2,6 @@ package com.onlydust.marketplace.indexer.domain.model.clean;
 
 import java.util.List;
 
-public record PullRequest(Integer id, User author, List<CodeReview> reviews, List<User> requestedReviewers,
+public record PullRequest(Long id, User author, List<CodeReview> reviews, List<User> requestedReviewers,
                           List<Commit> commits, List<CheckRun> checkRuns) {
 }
