+++ domain/src/main/java/com/onlydust/marketplace/indexer/domain/model/clean/User.java
@@ -4,5 +4,5 @@ import com.onlydust.marketplace.indexer.domain.model.raw.RawSocialAccount;
 
 import java.util.List;
 
-public record User(Integer id, String login, List<RawSocialAccount> socialAccounts) {
+public record User(Long id, String login, List<RawSocialAccount> socialAccounts) {
 }
